refactor: video page
