+++ webapp/src/components/account/_index.ts
@@ -1,0 +1,2 @@
+export { VideoAdminControls } from "./video-admin-controls";
+export type { IVideoAdminControls } from "./video-admin-controls";
