+++ webapp/src/components/comments/_index.ts
@@ -1,0 +1,2 @@
+export { NewCommentForm } from "./new-comment-form";
+export type { ICommentInit, INewCommentFormProps } from "./new-comment-form";
