+++ webapp/src/components/posts/_index.ts
@@ -1,0 +1,2 @@
+export { VideoView } from "./video";
+export type { IVideoViewProps } from "./video";
