+++ webapp/src/components/posts/video.tsx
@@ -1,0 +1,198 @@
+import { Tag, Avatar, Rate, Comment, List } from "antd";
+import {
+  faThumbsUp,
+  faUserCircle,
+} from "@fortawesome/free-solid-svg-icons";
+import { UserOutlined } from "@ant-design/icons";
+
+import { addComment } from "#api/comments";
+import { UserRank } from "#api/types";
+import { Icon } from "#components/icons";
+import { upvoteComment as apiUpvoteComment, postRating } from "#api/index";
+import { VideoPlayer } from "#components/video";
+import { LinkInternal } from "#components/links";
+import { VideoAdminControls } from "#components/account";
+import { NewCommentForm } from "#components/comments";
+
+export interface IVideoViewProps extends Record<string, unknown> {
+  id: number;
+  data: {
+    MPDLoc: string;
+    Title: string;
+    Views: string;
+    UploadDate: string;
+    Tags: string[];
+    AuthorID: number;
+    Username: string;
+    VideoDescription: string;
+    L: {
+      rank: number;
+    };
+  };
+  videoComments: {
+    upvote_count: number;
+    id: number;
+    user_has_upvoted: boolean;
+    content: string;
+    created: string;
+    fullname: string;
+  }[];
+
+  rating: number;
+  refreshComments: () => Promise<unknown>;
+  setRating: (rating: number) => void;
+  next_video: () => void;
+}
+
+export function VideoView(props: IVideoViewProps) {
+  let {
+    rating,
+    data,
+    id,
+    setRating,
+    next_video,
+    videoComments,
+    refreshComments,
+  } = props;
+
+  async function upvoteComment(commentID: number, has_upvoted: boolean) {
+    await apiUpvoteComment(commentID, !has_upvoted);
+    await refreshComments();
+  }
+
+  async function rate(rating: number) {
+    if (id == 0) {
+      // TODO: throw
+      return;
+    }
+    await postRating(id, rating);
+    setRating(rating);
+  }
+
+  // FIXME: new API endpoint
+  return (
+    <div className="bg-white dark:bg-black border">
+      <VideoPlayer url={data.MPDLoc} next_video={next_video} />
+      <div className="px-4 pt-4">
+        <div>
+          <span className="text-lg font-bold text-black dark:text-white">
+            {data.Title}
+          </span>
+          <span className="float-right">
+            <span className="text-black dark:text-white">
+              {data.Views} Views
+            </span>
+          </span>
+          <div className="inline-block relative top-5 float-right left-16 mr-2">
+            <Rate allowHalf={true} value={rating} onChange={rate}></Rate>
+          </div>
+          <br />
+          <span className="text-gray-600 text-xs">{data.UploadDate}</span>
+        </div>
+
+        <div className="my-2">
+          <span className="text-xs font-bold text-black dark:text-white mb-2">
+            Tags
+          </span>
+          <div className="border px-2 py-1">
+            {!data.Tags
+              ? "None"
+              : data.Tags.map((tag, idx) => {
+                  // TODO(ivan): add links to tags
+                  return (
+                    <div key={idx} className="my-1 inline-block">
+                      <LinkInternal href={`/?search=${tag}`}>
+                        <Tag color="blue">{tag}</Tag>
+                      </LinkInternal>
+                    </div>
+                  );
+                })}
+          </div>
+        </div>
+
+        <hr className="box-border w-full"></hr>
+        <div className="my-4">
+          <div>
+            <span className="h-20 w-20 inline-block">
+              <LinkInternal href={`/users/${data.AuthorID}`}>
+                <Avatar size={80} icon={<Icon icon={faUserCircle} />} />
+              </LinkInternal>
+            </span>
+            <span className="ml-2 pl-1 mt-2 inline-block align-top">
+              <LinkInternal href={`/users/${data.AuthorID}`}>
+                <b className="font-black text-blue-500 text-xl">
+                  {data.Username}
+                </b>
+              </LinkInternal>
+              <h1 className="text-black dark:text-white">0 subscribers</h1>
+            </span>
+          </div>
+          <div className="ml-20 pl-3 text-black dark:text-white">
+            {/* enjoy your XSS, bro */}
+            <span dangerouslySetInnerHTML={{ __html: data.VideoDescription }} />
+          </div>
+        </div>
+      </div>
+      {data.L && data.L.rank === UserRank.ADMIN && (
+        // @ts-expect-error types
+        <VideoAdminControls data={data} />
+      )}
+      <hr></hr>
+      <List
+        bordered={false}
+        split={false}
+        className="comment-list"
+        header={
+          <h2 className="ml-4 mb-0 text-xl text-black dark:text-white">
+            Comments ({videoComments.length})
+          </h2>
+        }
+        itemLayout="horizontal"
+        dataSource={videoComments}
+        renderItem={(item) => (
+          <li>
+            <Comment
+              className="border-0 text-black dark:text-white shadow-none"
+              actions={[
+                <span key={1} className="text-black dark:text-white text-bold">
+                  {item.upvote_count}
+                </span>,
+                <Icon
+                  key={2}
+                  onClick={() => upvoteComment(item.id, item.user_has_upvoted)}
+                  className={
+                    item.user_has_upvoted
+                      ? "max-h-4 mr-1 text-green-400"
+                      : "max-h-4 mr-1 text-gray-400"
+                  }
+                  icon={faThumbsUp}
+                />,
+              ]}
+              author={
+                <b className="text-black dark:text-white">{item.fullname}</b>
+              }
+              avatar={
+                <Avatar
+                  className="ml-4"
+                  shape="square"
+                  size={50}
+                  icon={<UserOutlined />}
+                />
+              }
+              content={item.content}
+              datetime={item.created}
+            />
+          </li>
+        )}
+      />
+      ,
+      <NewCommentForm
+        videoID={id}
+        onNewComment={async (commentInit) => {
+          await addComment(commentInit);
+          await refreshComments();
+        }}
+      />
+    </div>
+  );
+}
