+++ webapp/src/components/video/_index.ts
@@ -1,0 +1,2 @@
+export { VideoPlayer } from "./player";
+export type { IVideoPlayerProps } from "./player";
