+++ webapp/src/pages/videos/[video_id].tsx
@@ -1,42 +1,10 @@
-import { useCallback, useEffect, useRef, useState } from "react";
-import { Tag, Avatar, Button, Input, Rate, Comment, List } from "antd";
-import Link from "next/link";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faThumbsUp,
-  faThumbsDown,
-  faUserCircle,
-  faUser,
-} from "@fortawesome/free-solid-svg-icons";
-import { useFormik } from "formik";
-import { UserOutlined } from "@ant-design/icons";
+import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
-import type { SyntheticEvent } from "react";
 
-import {
-  getComments,
-  getVideo,
-  getUserdata,
-  deleteVideo as apiDeleteVideo,
-  approveVideo as apiApproveVideo,
-  postComment,
-  upvoteComment as apiUpvoteComment,
-  postRating,
-} from "#api/index";
-import { UserRank } from "#api/types";
+import { getComments, getVideo, getUserdata } from "#api/index";
 import { VideoList } from "#components/video-list";
 import { Header } from "#components/header";
-import { LinkExternal, LinkInternal } from "#components/links";
-import { Icon } from "#components/icons";
-import {
-  FormClient,
-  type IFormElements,
-  type ISubmitEvent,
-} from "#components/forms";
-import { Hidden, Text } from "#components/inputs";
-
-const VIDEO_WIDTH = 44;
-const VIDEO_HEIGHT = (9 / 16) * VIDEO_WIDTH;
+import { VideoView } from "#components/posts";
 
 interface IPageData extends Record<string, unknown> {
   RecommendedVideos: Record<string, unknown>[];
@@ -115,403 +83,6 @@ function VideosPage() {
         </div>
       </div>
     </>
-  );
-}
-
-interface IVideoPlayerProps extends Record<string, unknown> {
-  url: string;
-  next_video: (event: SyntheticEvent<HTMLVideoElement, Event>) => void;
-}
-
-function VideoPlayer(props: IVideoPlayerProps) {
-  let { url, next_video } = props;
-  let videoRef = useRef<HTMLVideoElement>(null);
-  var url_without_mpd = url.slice(0, -4);
-
-  useEffect(() => {
-    let video = videoRef.current;
-    if (video == null) return;
-    video.setAttribute("src", url_without_mpd);
-    video.load();
-    video.play();
-  }, [videoRef, url]);
-
-  return (
-    <>
-      <video
-        ref={videoRef}
-        id="my-player"
-        className="bg-black dark:bg-black w-full max-w-screen-lg object-contain object-center z-0"
-        style={{ height: `${VIDEO_HEIGHT}rem` }}
-        onEnded={next_video}
-        controls
-        preload="auto"
-        data-setup="{}"
-      >
-        <source src={url_without_mpd} type="video/mp4"></source>
-
-        <p className="vjs-no-js">
-          To view this video please enable JavaScript, and consider upgrading to
-          a web browser that
-          <LinkExternal
-            href="https://videojs.com/html5-video-support/"
-            target="_blank"
-            rel="noreferrer"
-          >
-            supports HTML5 video
-          </LinkExternal>
-        </p>
-      </video>
-    </>
-  );
-}
-
-interface IVideoAdminControls extends Record<string, unknown> {
-  data: {
-    VideoID: number;
-  };
-}
-
-function VideoAdminControls(props: IVideoAdminControls) {
-  const router = useRouter();
-  let { data } = props;
-  let [approvedVideo, setApprovedVideo] = useState(false);
-  let approvingVideo = useRef(false);
-  let deletingVideo = useRef(false);
-
-  let deleteVideo = () => {
-    if (deletingVideo.current) return;
-
-    deletingVideo.current = true;
-
-    let run = async () => {
-      await apiDeleteVideo(data.VideoID);
-      deletingVideo.current = false;
-      router.push("/");
-    };
-    run();
-    // TODO: error future handler
-  };
-
-  let approveVideo = useCallback(() => {
-    if (approvingVideo.current) return;
-
-    let run = async () => {
-      await apiApproveVideo(data.VideoID);
-      setApprovedVideo(true);
-    };
-    approvingVideo.current = true;
-
-    run().finally(() => {
-      approvingVideo.current = false;
-    });
-    // TODO: error future handler
-  }, [data, approvingVideo, setApprovedVideo]);
-
-  return (
-    <>
-      <hr />
-      <div className="my-4">
-        <b className="text-2xl text-black dark:text-white ml-4">
-          Admin controls:
-        </b>
-        <Button
-          type="primary"
-          className="mx-4"
-          disabled={approvedVideo}
-          onClick={approveVideo}
-        >
-          {approvedVideo ? "Approved" : "Approve"}
-        </Button>
-        <Button type="primary" className="mx-4" onClick={deleteVideo}>
-          Delete video
-        </Button>
-      </div>
-    </>
-  );
-}
-
-interface IVideoViewProps extends Record<string, unknown> {
-  id: number;
-  data: {
-    MPDLoc: string;
-    Title: string;
-    Views: string;
-    UploadDate: string;
-    Tags: string[];
-    AuthorID: number;
-    Username: string;
-    VideoDescription: string;
-    L: {
-      rank: number;
-    };
-  };
-  videoComments: {
-    upvote_count: number;
-    id: number;
-    user_has_upvoted: boolean;
-    content: string;
-    created: string;
-    fullname: string;
-  }[];
-
-  rating: number;
-  refreshComments: () => Promise<unknown>;
-  setRating: (rating: number) => void;
-  next_video: () => void;
-}
-
-function VideoView(props: IVideoViewProps) {
-  let {
-    rating,
-    data,
-    id,
-    setRating,
-    next_video,
-    videoComments,
-    refreshComments,
-  } = props;
-
-  // video_id, content (content of comment), and parent (parent comment id if a reply)
-  let formik = useFormik({
-    initialValues: {
-      content: "",
-      video_id: id,
-      parent: "",
-    },
-    enableReinitialize: true,
-    onSubmit: async (values) => {
-      await postComment(values);
-      await refreshComments();
-    },
-  });
-
-  async function upvoteComment(commentID: number, has_upvoted: boolean) {
-    await apiUpvoteComment(commentID, !has_upvoted);
-    await refreshComments();
-  }
-
-  async function rate(rating: number) {
-    if (id == 0) {
-      // TODO: throw
-      return;
-    }
-    await postRating(id, rating);
-    setRating(rating);
-  }
-
-  // FIXME: new API endpoint
-  return (
-    <div className="bg-white dark:bg-black border">
-      <VideoPlayer url={data.MPDLoc} next_video={next_video} />
-      <div className="px-4 pt-4">
-        <div>
-          <span className="text-lg font-bold text-black dark:text-white">
-            {data.Title}
-          </span>
-          <span className="float-right">
-            <span className="text-black dark:text-white">
-              {data.Views} Views
-            </span>
-          </span>
-          <div className="inline-block relative top-5 float-right left-16 mr-2">
-            <Rate allowHalf={true} value={rating} onChange={rate}></Rate>
-          </div>
-          <br />
-          <span className="text-gray-600 text-xs">{data.UploadDate}</span>
-        </div>
-
-        <div className="my-2">
-          <span className="text-xs font-bold text-black dark:text-white mb-2">
-            Tags
-          </span>
-          <div className="border px-2 py-1">
-            {!data.Tags
-              ? "None"
-              : data.Tags.map((tag, idx) => {
-                  // TODO(ivan): add links to tags
-                  return (
-                    <div key={idx} className="my-1 inline-block">
-                      <Link href={`/?search=${tag}`}>
-                        <Tag color="blue">{tag}</Tag>
-                      </Link>
-                    </div>
-                  );
-                })}
-          </div>
-        </div>
-
-        <hr className="box-border w-full"></hr>
-        <div className="my-4">
-          <div>
-            <span className="h-20 w-20 inline-block">
-              <LinkInternal href={`/users/${data.AuthorID}`}>
-                <Avatar size={80} icon={<Icon icon={faUserCircle} />} />
-              </LinkInternal>
-            </span>
-            <span className="ml-2 pl-1 mt-2 inline-block align-top">
-              <LinkInternal href={`/users/${data.AuthorID}`}>
-                <b className="font-black text-blue-500 text-xl">
-                  {data.Username}
-                </b>
-              </LinkInternal>
-              <h1 className="text-black dark:text-white">0 subscribers</h1>
-            </span>
-          </div>
-          <div className="ml-20 pl-3 text-black dark:text-white">
-            {/* enjoy your XSS, bro */}
-            <span dangerouslySetInnerHTML={{ __html: data.VideoDescription }} />
-          </div>
-        </div>
-      </div>
-      {data.L && data.L.rank === UserRank.ADMIN && (
-        // @ts-expect-error types
-        <VideoAdminControls data={data}></VideoAdminControls>
-      )}
-      <hr></hr>
-      <List
-        bordered={false}
-        split={false}
-        className="comment-list"
-        header={
-          <h2 className="ml-4 mb-0 text-xl text-black dark:text-white">
-            Comments ({videoComments.length})
-          </h2>
-        }
-        itemLayout="horizontal"
-        dataSource={videoComments}
-        renderItem={(item) => (
-          <li>
-            <Comment
-              className="border-0 text-black dark:text-white shadow-none"
-              actions={[
-                <span key={1} className="text-black dark:text-white text-bold">
-                  {item.upvote_count}
-                </span>,
-                <Icon
-                  key={2}
-                  onClick={() => upvoteComment(item.id, item.user_has_upvoted)}
-                  className={
-                    item.user_has_upvoted
-                      ? "max-h-4 mr-1 text-green-400"
-                      : "max-h-4 mr-1 text-gray-400"
-                  }
-                  icon={faThumbsUp}
-                />,
-              ]}
-              author={
-                <b className="text-black dark:text-white">{item.fullname}</b>
-              }
-              avatar={
-                <Avatar
-                  className="ml-4"
-                  shape="square"
-                  size={50}
-                  icon={<UserOutlined />}
-                />
-              }
-              content={item.content}
-              datetime={item.created}
-            />
-          </li>
-        )}
-      />
-      ,
-      <NewCommentForm videoID={id} />
-      <form onSubmit={formik.handleSubmit}>
-        <Input.Group>
-          <Input
-            name="content"
-            // @ts-expect-error types
-            values={formik.values.content}
-            onChange={formik.handleChange}
-            size="large"
-            placeholder="<your comment here>"
-            prefix={
-              <FontAwesomeIcon
-                className="max-h-4 mr-1 text-gray-400"
-                icon={faUser}
-              />
-            }
-          />
-        </Input.Group>
-        <Input.Group>
-          <Button block type="primary" htmlType="submit" size="large">
-            Submit
-          </Button>
-        </Input.Group>
-      </form>
-    </div>
-  );
-}
-
-export interface ICommentInit {
-  video_id: string;
-  content: string;
-  parent: string;
-}
-
-const FIELD_NAMES = {
-  MESSAGE: "content",
-  VIDEO_ID: "video_id",
-  PARENT: "parent",
-} as const;
-type IFieldName = typeof FIELD_NAMES[keyof typeof FIELD_NAMES];
-
-interface INewCommentFormProps {
-  videoID: number;
-  parentID?: string;
-
-  onNewComment: (commentInit: ICommentInit) => Promise<void>;
-}
-
-function NewCommentForm({
-  videoID,
-  parentID,
-  onNewComment,
-}: INewCommentFormProps) {
-  async function handleSubmit(event: ISubmitEvent) {
-    const fields = event.currentTarget.elements as IFormElements<IFieldName>;
-    const commentInit = Object.values(FIELD_NAMES).reduce(
-      (commentInit, fieldName) => {
-        switch (fieldName) {
-          case FIELD_NAMES.MESSAGE:
-          case FIELD_NAMES.PARENT:
-          case FIELD_NAMES.VIDEO_ID: {
-            const fieldElement = fields[fieldName];
-            commentInit[fieldName] = fieldElement.value
-            break;
-          }
-
-          default:
-            throw new Error(
-              `The field "${fieldName}" is missing from the form.`
-            );
-        }
-        return commentInit;
-      },
-      {} as ICommentInit
-    );
-
-    await onNewComment(commentInit);
-  }
-
-  return (
-    <FormClient onSubmit={handleSubmit} id="new-comment">
-      <Hidden
-        id="new-comment-parent-id"
-        name={FIELD_NAMES.PARENT}
-        defaultValue={parentID}
-      />
-      <Hidden
-        id="new-comment-video-id"
-        name={FIELD_NAMES.VIDEO_ID}
-        defaultValue={videoID}
-      />
-      <Text id="new-comment-content" name={FIELD_NAMES.MESSAGE}>
-        New comment
-      </Text>
-    </FormClient>
   );
 }
